Reject no-authorization-specified when signing
GetAndValidateDigest treated no digest specified as Digest::NONE for
purposes of implementing the legacy keymaster interface. However, on new
Keymaster interface we need to return UNSUPPORTED_DIGEST in some cases
when no digest is specified (as exactly 1 is mandated).

Previously in these case, we would return INCOMPATIBLE_DIGEST. This
change brings the reference implementation in line with the spec and
VTS.

Bug: 129294086
Test: VTS 3, VTS 4 against reference
Change-Id: Ib658fb6fc3384320c90989669824e5270e025937