Fixing up reference impl to match new CDDL
This changes the reference implementation to add the MAC tag from the
MACing operation over the list of public keys to be signed into the AAD
for the signature of the ephemeral MAC key. This prevents a potential
vulnerability that would arise whereby an attacker could get arbitrary
key pairs signed if they compromised the GEEK private key.

Bug: 189018262
Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: I31f352667e50d273bd1d545ee786ce8b09d4eb4d