Add production GEEK validation code
Add the 2021 production Google Endpoint Encryption Key to the EEK
validation code.

The GEEK is a full COSE_Key, containing not only the pubkey, but
key type, algorithm, curve, and public key all in one blob.

Also switch to using string_view instead of vector so that we're not
using non-trivially constructed types in static variables. This allows
us to make the values constexpr, avoiding all heap use.

Lastly, compare both the pubkey and curve. This way we are certain
that we're not confusing one key type for another.

Test: libkeymint_remote_prov_support_test
Bug: 191301285
Change-Id: Id90a5206a9a77dfed8547b2e7444a1fedd2edf18
Merged-In: Id90a5206a9a77dfed8547b2e7444a1fedd2edf18