Breaking out mac key signing into the context
This change consolidates the functionality around signing the
MAC key and fetching the Boot Certificate Chain into one
function. These are the two parts to creating a
ProtectedData payload. It makes sense to consolidate the
two functions into one and just return the entire payload, and
it simplifies the testing implementation on the hardware backend.

Bug: 194509629
Test: atest VtsHalRemotelyProvisionedComponentTest
Change-Id: I60d2db26e71e88c268ce4d325ce4132886956088