Clean up some compiler warnings
NOLINT was on the wrong line of code in remote_provisioning_utils.h

C++14 does not appear to be fully supported everywhere that
android_keymaster_messages.h is included, which results in build spam
around the C++14 use of a constexpr function. Remove constexpr for now
and just rely on the compiler to do the optimization.

Test: keymaster_tests
Change-Id: Id9db41700fbb796c823b90f686cc60095c6adc39